src/requirements_win7.txt
pyside6-essentials==6.1.3
requests==2.26.0
urllib3==1.25.11
pillow==8.3.2
bs4==0.0.1
lxml==4.6.4
pycryptodomex==3.12.0
Pysocks==1.7.1
natsort==8.2.0
curl_cffi==0.5.10
jmcomic>=2.4.3
webdavclient3==3.14.6
tqdm==4.66.4
pysmb==1.2.9.1